DAX-534 Only trigger shadow customer creation if quote is guest
Under certain conditions, i.e. if customer e-mail was changed after the
quote was created, a shadow customer was created for the old e-mail
adress on the quote, even though the quote was not a guest quote, and
was linked to an existing customer.

This shadow customer would then trip up certain validations done on the
quote, and cause the order submission to fail.